Add class 'Calendar' to XCalendar.Core and implemented it into CalendarView
