--- XCalendar.Forms/Enums/DayState.cs
@@ -1,11 +1,0 @@
-﻿namespace XCalendar.Forms.Enums
-{
-    public enum DayState
-    {
-        CurrentMonth,
-        OtherMonth,
-        Today,
-        Selected,
-        Invalid
-    }
-}
