--- XCalendar.Forms/Enums/NavigationTimeUnit.cs
@@ -1,10 +1,0 @@
-﻿namespace XCalendar.Forms.Enums
-{
-    public enum NavigationTimeUnit
-    {
-        Day,
-        Week,
-        Month,
-        Year
-    }
-}
