--- XCalendar.Forms/Enums/SelectionAction.cs
@@ -1,10 +1,0 @@
-﻿namespace XCalendar.Forms.Enums
-{
-    public enum SelectionAction
-    {
-        Add,
-        Remove,
-        Modify,
-        Replace
-    }
-}
