--- XCalendar.Forms/Enums/SelectionType.cs
@@ -1,9 +1,0 @@
-﻿namespace XCalendar.Forms.Enums
-{
-    public enum SelectionType
-    {
-        None,
-        Single,
-        Range
-    }
-}
