--- XCalendar.Maui/Enums/PageStartMode.cs
@@ -1,9 +1,0 @@
-﻿namespace XCalendar.Maui.Enums
-{
-    public enum PageStartMode
-    {
-        FirstDayOfWeek,
-        FirstDayOfMonth,
-        FirstDayOfYear
-    }
-}
