+++ XCalendar.Maui/Views/CalendarView.xaml
@@ -56,7 +56,7 @@
                     WidthRequest="30">
 
                     <xc:BorderedLabel.GestureRecognizers>
-                        <TapGestureRecognizer Command="{Binding NavigateCalendarCommand, Source={x:Reference MainCalendarView}}" CommandParameter="{Binding BackwardsNavigationAmount, Source={x:Reference MainCalendarView}}"/>
+                        <TapGestureRecognizer Command="{Binding NavigateCalendarCommand, Source={x:Reference MainCalendarView}}" CommandParameter="{Binding Calendar.BackwardsNavigationAmount, Source={x:Reference MainCalendarView}}"/>
                     </xc:BorderedLabel.GestureRecognizers>
 
                 </xc:BorderedLabel>
@@ -65,7 +65,7 @@
                     FontAttributes="Bold"
                     FontSize="18"
                     HorizontalTextAlignment="Center"
-                    Text="{Binding NavigatedDate, StringFormat='{0:MMMM yyy}', Source={x:Reference MainCalendarView}}"
+                    Text="{Binding Calendar.NavigatedDate, StringFormat='{0:MMMM yyy}', Source={x:Reference MainCalendarView}}"
                     TextColor="{Binding NavigationTextColor, Source={x:Reference MainCalendarView}}"
                     VerticalTextAlignment="Center"/>
 
@@ -83,7 +83,7 @@
                     WidthRequest="30">
 
                     <xc:BorderedLabel.GestureRecognizers>
-                        <TapGestureRecognizer Command="{Binding NavigateCalendarCommand, Source={x:Reference MainCalendarView}}" CommandParameter="{Binding ForwardsNavigationAmount, Source={x:Reference MainCalendarView}}"/>
+                        <TapGestureRecognizer Command="{Binding NavigateCalendarCommand, Source={x:Reference MainCalendarView}}" CommandParameter="{Binding Calendar.ForwardsNavigationAmount, Source={x:Reference MainCalendarView}}"/>
                     </xc:BorderedLabel.GestureRecognizers>
                     
                 </xc:BorderedLabel>
@@ -97,7 +97,7 @@
             <CollectionView
                 Margin="0,5,0,0"
                 HeightRequest="{Binding DayNamesHeightRequest, Source={x:Reference MainCalendarView}}"
-                ItemsSource="{Binding DayNamesOrder, Source={x:Reference MainCalendarView}}">
+                ItemsSource="{Binding Calendar.DayNamesOrder, Source={x:Reference MainCalendarView}}">
 
                 <CollectionView.Style>
                     <Style TargetType="{x:Type CollectionView}">
@@ -128,7 +128,7 @@
                     <GridItemsLayout
                         HorizontalItemSpacing="{Binding DayNameHorizontalSpacing, Source={x:Reference MainCalendarView}}"
                         Orientation="Vertical"
-                        Span="{Binding DayNamesOrder.Count, Source={x:Reference MainCalendarView}}"
+                        Span="{Binding Calendar.DayNamesOrder.Count, Source={x:Reference MainCalendarView}}"
                         VerticalItemSpacing="{Binding DayNameVerticalSpacing, Source={x:Reference MainCalendarView}}"/>
                 </CollectionView.ItemsLayout>
 
@@ -138,7 +138,7 @@
 
         <!--#region MonthView-->
         <ContentView ControlTemplate="{Binding MonthViewTemplate, Source={x:Reference MainCalendarView}}">
-            <CollectionView HeightRequest="{Binding MonthViewHeightRequest, Source={x:Reference MainCalendarView}}" ItemsSource="{Binding Days, Source={x:Reference MainCalendarView}}">
+            <CollectionView HeightRequest="{Binding MonthViewHeightRequest, Source={x:Reference MainCalendarView}}" ItemsSource="{Binding Calendar.Days, Source={x:Reference MainCalendarView}}">
                 <CollectionView.Style>
                     <Style TargetType="{x:Type CollectionView}">
                         <Setter Property="ItemTemplate">
@@ -167,7 +167,7 @@
                 </CollectionView.Style>
 
                 <CollectionView.ItemsLayout>
-                    <GridItemsLayout Orientation="Vertical" Span="{Binding DayNamesOrder.Count, Source={x:Reference MainCalendarView}}"/>
+                    <GridItemsLayout Orientation="Vertical" Span="{Binding Calendar.DayNamesOrder.Count, Source={x:Reference MainCalendarView}}"/>
                 </CollectionView.ItemsLayout>
 
             </CollectionView>
