+++ XCalendarFormsSample/XCalendarFormsSample/Helpers/PopupHelper.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xamarin.CommunityToolkit.Extensions;
 using Xamarin.Forms;
-using XCalendar.Forms.Enums;
+using XCalendar.Core.Enums;
 using XCalendarFormsSample.Popups;
 
 namespace XCalendarFormsSample.Helpers
