+++ XCalendarFormsSample/XCalendarFormsSample/Popups/DatePickerDialogPopup.xaml
@@ -8,9 +8,11 @@
     xmlns:xcConverters="clr-namespace:XCalendar.Forms.Converters;assembly=XCalendar.Forms"
     xmlns:xcModels="clr-namespace:XCalendar.Core.Models;assembly=XCalendar.Core"
     xmlns:xct="http://xamarin.com/schemas/2020/toolkit"
+    xmlns:Popups="clr-namespace:XCalendarFormsSample.Popups"
     x:Name="This"
     x:TypeArguments="System:DateTime"
-    Size="320, 475">
+    Size="320, 475"
+    x:DataType="{x:Type Popups:DatePickerDialogPopup}">
 
     <StackLayout BackgroundColor="White" Spacing="0">
 
@@ -47,13 +49,10 @@
             Margin="0,5,0,0"
             Padding="20,0,20,0"
             MonthViewHeightRequest="246"
-            NavigatedDate="{Binding NavigatedDate, Source={x:Reference This}}"
+            Calendar="{Binding Calendar, Source={x:Reference This}}"
             NavigationArrowColor="{StaticResource ContentTextColor}"
             NavigationBackgroundColor="Transparent"
-            NavigationTextColor="{StaticResource ContentTextColor}"
-            SelectedDates="{Binding SelectedDates, Source={x:Reference This}}"
-            SelectionAction="Replace"
-            SelectionType="Single">
+            NavigationTextColor="{StaticResource ContentTextColor}">
 
             <xc:CalendarView.DayNameTemplate>
                 <DataTemplate x:DataType="{x:Type System:DayOfWeek}">
