+++ XCalendarFormsSample/XCalendarFormsSample/Views/SelectionExamplePage.xaml
@@ -21,7 +21,7 @@
                 </Style>
             </StackLayout.Resources>
 
-            <xc:CalendarView x:Name="MainCalendarView" Style="{StaticResource DefaultCalendarViewStyle}" SelectionAction="{Binding SelectionAction}" SelectionType="{Binding SelectionType}">
+            <xc:CalendarView x:Name="MainCalendarView" Style="{StaticResource DefaultCalendarViewStyle}" Calendar="{Binding Calendar}">
                 
                 <!--Not Required, used only for styling.-->
                 <xc:CalendarView.DayTemplate>
@@ -49,7 +49,7 @@
                     xct:TouchEffect.NativeAnimation="True"
                     FontSize="{StaticResource SmallFontSize}"
                     HorizontalTextAlignment="Center"
-                    Text="{Binding SelectionType}"
+                    Text="{Binding Calendar.SelectionType}"
                     TextColor="{StaticResource TappableSettingTextColor}"
                     VerticalTextAlignment="Center"/>
             </Grid>
@@ -67,7 +67,7 @@
                     xct:TouchEffect.NativeAnimation="True"
                     FontSize="{StaticResource SmallFontSize}"
                     HorizontalTextAlignment="Center"
-                    Text="{Binding SelectionAction}"
+                    Text="{Binding Calendar.SelectionAction}"
                     TextColor="{StaticResource TappableSettingTextColor}"
                     VerticalTextAlignment="Center"/>
             </Grid>
