+++ XCalendarFormsSample/XCalendarFormsSample/Views/UsingCalendarDayViewExamplePage.xaml
@@ -12,7 +12,7 @@
     <ScrollView>
         <StackLayout Spacing="0">
 
-            <xc:CalendarView Style="{StaticResource DefaultCalendarViewStyle}" MonthViewHeightRequest="320" x:Name="MainCalendarView" NavigatedDate="{Binding NavigatedDate}" SelectionType="Single" SelectionAction="Replace" SelectedDates="{Binding SelectedDates}">
+            <xc:CalendarView Style="{StaticResource DefaultCalendarViewStyle}" MonthViewHeightRequest="320" x:Name="MainCalendarView" Calendar="{Binding Calendar}">
                 <xc:CalendarView.DayTemplate>
                     <DataTemplate x:DataType="{x:Type xcInterfaces:ICalendarDay}">
                         <ContentView>
@@ -39,9 +39,9 @@
                     HeightRequest="47"
                     Margin="0,5,0,5"
                     CalendarView="{Binding ., Source={x:Reference MainCalendarView}}"
-                    DateTime="{Binding NavigatedDate}"/>
-                
-                <DatePicker Format="/MM/yyy" Date="{Binding NavigatedDate}"/>
+                    DateTime="{Binding Calendar.NavigatedDate}"/>
+
+                <DatePicker Format="/MM/yyy" Date="{Binding Calendar.NavigatedDate}"/>
                 
             </StackLayout>
