+++ XCalendarMauiSample/ViewModels/EventCalendarExampleViewModel.cs
@@ -1,5 +1,9 @@
-﻿using System.Collections.Specialized;
-using XCalendar.Maui;
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Linq;
+using XCalendar.Core.Enums;
+using XCalendar.Core.Models;
 using XCalendarMauiSample.Models;
 
 namespace XCalendarMauiSample.ViewModels
@@ -7,9 +11,15 @@ namespace XCalendarMauiSample.ViewModels
     public class EventCalendarExampleViewModel : BaseViewModel
     {
         #region Properties
+        public Calendar Calendar { get; set; } = new Calendar()
+        {
+            SelectedDates = new ObservableRangeCollection<DateTime>(),
+            SelectionAction = SelectionAction.Modify,
+            SelectionType = SelectionType.Single
+        };
         public static readonly Random Random = new Random();
         public List<Color> Colors { get; } = new List<Color>() { Microsoft.Maui.Graphics.Colors.Red, Microsoft.Maui.Graphics.Colors.Orange, Microsoft.Maui.Graphics.Colors.Yellow, Color.FromArgb("#00A000"), Microsoft.Maui.Graphics.Colors.Blue, Color.FromArgb("#8010E0") };
-        public EventDayResolver EventDayResolver { get; set; } = new EventDayResolver();
+        public EventDayResolver EventDayResolver { get; } = new EventDayResolver();
         public ObservableRangeCollection<Event> Events { get; } = new ObservableRangeCollection<Event>()
         {
             new Event() { Title = "Bowling", Description = "Bowling with friends" },
@@ -47,7 +57,6 @@ namespace XCalendarMauiSample.ViewModels
             new Event() { Title = "Sailing", Description = "Sailing with friends" },
             new Event() { Title = "Cooking", Description = "Cooking with friends" }
         };
-        public ObservableRangeCollection<DateTime> SelectedDates { get; } = new ObservableRangeCollection<DateTime>();
         public ObservableRangeCollection<Event> SelectedEvents { get; } = new ObservableRangeCollection<Event>();
         #endregion
 
@@ -62,14 +71,15 @@ namespace XCalendarMauiSample.ViewModels
 
             EventDayResolver.Events = Events;
 
-            SelectedDates.CollectionChanged += SelectedDates_CollectionChanged;
+            Calendar.DayResolver = EventDayResolver;
+            Calendar.SelectedDates.CollectionChanged += SelectedDates_CollectionChanged;
         }
         #endregion
 
         #region Methods
         private void SelectedDates_CollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
         {
-            SelectedEvents.ReplaceRange(Events.Where(x => SelectedDates.Any(y => x.DateTime.Date == y.Date)).OrderByDescending(x => x.DateTime));
+            SelectedEvents.ReplaceRange(Events.Where(x => Calendar.SelectedDates.Any(y => x.DateTime.Date == y.Date)).OrderByDescending(x => x.DateTime));
         }
         #endregion
     }
