+++ XCalendarMauiSample/ViewModels/ExamplesViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using PropertyChanged;
 using System.Windows.Input;
+using XCalendar.Core.Models;
 using XCalendar.Maui;
 using XCalendar.Maui.Views;
 using XCalendarMauiSample.Models;
