+++ XCalendarMauiSample/Views/EventCalendarExamplePage.xaml
@@ -37,14 +37,11 @@
                 x:Name="MainCalendarView"
                 Style="{StaticResource DefaultCalendarViewStyle}"
                 Grid.Row="0"
-                DayResolver="{Binding EventDayResolver}"
                 DayNameTextColor="{StaticResource ContentTextColor}"
                 NavigationArrowColor="{StaticResource ContentTextColor}"
                 NavigationBackgroundColor="Transparent"
                 NavigationTextColor="{StaticResource ContentTextColor}"
-                SelectedDates="{Binding SelectedDates}"
-                SelectionAction="Modify"
-                SelectionType="Single">
+                Calendar="{Binding Calendar}">
 
                 <xc:CalendarView.DayTemplate>
                     <DataTemplate x:DataType="{x:Type Models:EventDay}">
